Load megamenus after loading body and cleanup header code
