+++ scripts/scripts.js
@@ -15,6 +15,8 @@ import {
 } from './lib-franklin.js';
 import TEMPLATE_LIST from '../templates/config.js';
 
+import fetchAndStyleHeaderMegamenus from '../blocks/header/header-megamenu.js';
+
 const LCP_BLOCKS = []; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'project-1'; // add your RUM generation information here
 
@@ -117,7 +119,12 @@ export function addFavIcon(href, rel = 'icon') {
  */
 async function loadLazy(doc) {
   const main = doc.querySelector('main');
+
   loadHeader(doc.querySelector('header'));
+
+  // Fetch and style the megamenus in the background
+  await fetchAndStyleHeaderMegamenus(doc.querySelector('header'));
+
   await loadBlocks(main);
 
   const { hash } = window.location;
